% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{recode_landcover_codes}
\alias{recode_landcover_codes}
\title{Recode a land-cover code vector to names}
\usage{
recode_landcover_codes(x, key, as_factor = TRUE)
}
\arguments{
\item{x}{Integer/character vector of land-cover codes (e.g., 1..8).}

\item{key}{Named character vector mapping codes -> labels
(e.g., c("1"="urban", "2"="croplands", ...)).}

\item{as_factor}{Logical; return a factor with levels in key order? Default TRUE.}
}
\value{
A character (or factor) vector of labels; unknown codes become NA.
}
\description{
Recode a land-cover code vector to names
}
\examples{
landcover_key <- c(
  "1"="urban","2"="croplands","3"="forest","4"="shrub & scrub",
  "5"="grass","6"="water","7"="wetlands","8"="bare surface"
)
x <- c(1,3,6,NA,9)
recode_landcover_codes(x, landcover_key)
}
